refactor: CarUtil ->  Move  피드백 주신걸 듣고  랜덤과 무브로 나눌려 생각을 했다 Move (이동과 관련된 함수들) 변경해보았습니다.
